hot-fix: modularize the dashboard
